Add files via upload
Updated Tic-tac-toe assignment. These files contain working cell selection and console.log of when a winner is chosen. I still need to create a reset function that resets the board and a function to alert the winners if they won or if there has been a draw between the 2 players.